LPD-85801 SF

+++ modules/apps/address/address-test/src/testIntegration/java/com/liferay/address/internal/upgrade/v1_0_0/test/CountryUpgradeProcessTest.java
@@ -139,7 +139,7 @@ public class CountryUpgradeProcessTest {
 			try (ResultSet resultSet = preparedStatement.executeQuery()) {
 				resultSet.next();
 
-				return resultSet.getInt("count");
+				return resultSet.getLong("count");
 			}
 		}
 	}
