+++ modules/apps/portal/portal-db-partition-test/src/testIntegration/java/com/liferay/portal/db/partition/test/DBPartitionTest.java
@@ -269,7 +269,7 @@ public class DBPartitionTest extends BaseDBPartitionTestCase {
 					ResultSet resultSet = preparedStatement.executeQuery()) {
 
 					if (resultSet.next()) {
-						rowCount = resultSet.getInt("count");
+						rowCount = resultSet.getLong("count");
 					}
 				}
